moved gimmeCode back into main switch - sorry!
